FinGuard/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
google-cloud-firestore==2.16.0
google-cloud-bigquery==3.25.0
vertexai==1.69.0
twilio==9.1.0
pydantic==2.8.2
python-dotenv==1.0.1